+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+master (XXXX-XX-XX)
+--------------------
+* Feature: Add support for the optional 'networks' parameter on server creation.
+
 1.6.3 (2020-01-09)
 --------------------